rename emulator.cpp to main.cpp and emulation.cpp to emulator.cpp
